replace event module by the cert module
