+++ src/lib.rs
@@ -1,3 +1,3 @@
-pub mod event;
+pub mod cert;
 pub mod sftp;
 pub mod sql;
